--- app/lib/desktop.ini
@@ -1,2 +1,0 @@
-[LocalizedFileNames]
-mysql-connector-java-8.0.18.jar=@mysql-connector-java-8.0.18.jar,0
