+++ app/nbproject/project.properties
@@ -5,6 +5,7 @@ annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
 application.title=app
 application.vendor=ymegam
+auxiliary.org-netbeans-spi-editor-hints-projects.perProjectHintSettingsFile=nbproject/cfg_hints.xml
 build.classes.dir=${build.dir}/classes
 build.classes.excludes=**/*.java,**/*.form
 # This directory is removed when the project is cleaned:
@@ -33,14 +34,14 @@ dist.jar=${dist.dir}/app.jar
 dist.javadoc.dir=${dist.dir}/javadoc
 endorsed.classpath=
 excludes=
-file.reference.mysql-connector-java-8.0.18.jar=lib/mysql-connector-java-8.0.18.jar
+file.reference.java.app-lib=lib
 includes=**
 # Non-JavaFX jar file creation is deactivated in JavaFX 2.0+ projects
 jar.archive.disabled=true
 jar.compress=false
 javac.classpath=\
-    ${javafx.classpath.extension}:\
-    ${file.reference.mysql-connector-java-8.0.18.jar}
+    ${file.reference.java.app-lib}:\
+    ${javafx.classpath.extension}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
@@ -76,11 +77,16 @@ javafx.classpath.extension=\
     ${java.home}/lib/javaws.jar:\
     ${java.home}/lib/deploy.jar:\
     ${java.home}/lib/plugin.jar
+javafx.deploy.adddesktopshortcut=false
+javafx.deploy.addstartmenushortcut=false
 javafx.deploy.allowoffline=true
 # If true, application update mode is set to 'background', if false, update mode is set to 'eager'
 javafx.deploy.backgroundupdate=false
+javafx.deploy.disable.proxy=false
 javafx.deploy.embedJNLP=true
 javafx.deploy.includeDT=true
+javafx.deploy.installpermanently=false
+javafx.deploy.permissionselevated=false
 # Set true to prevent creation of temporary copy of deployment artifacts before each run (disables concurrent runs)
 javafx.disable.concurrent.runs=false
 # Set true to enable multiple concurrent runs of the same WebStart or Run-in-Browser project
@@ -89,7 +95,7 @@ javafx.enable.concurrent.external.runs=false
 javafx.enabled=true
 javafx.fallback.class=com.javafx.main.NoJavaFXFallback
 # Main class for JavaFX
-javafx.main.class=app.MainApp
+javafx.main.class=app.inicio.MainApp
 javafx.preloader.class=
 # This project does not use Preloader
 javafx.preloader.enabled=false
@@ -101,6 +107,9 @@ javafx.preloader.type=none
 javafx.rebase.libs=false
 javafx.run.height=600
 javafx.run.width=800
+javafx.signing.blob=false
+javafx.signing.enabled=false
+javafx.signing.type=notsigned
 jlink.launcher=false
 jlink.launcher.name=app
 # Pre-JavaFX 2.0 WebStart is deactivated in JavaFX 2.0+ projects
@@ -114,6 +123,7 @@ manifest.custom.permissions=
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
+native.bundling.enabled=false
 platform.active=default_platform
 run.classpath=\
     ${dist.jar}:\
@@ -127,5 +137,6 @@ run.test.classpath=\
 run.test.modulepath=\
     ${javac.test.modulepath}
 source.encoding=UTF-8
-src.src.dir=src
+src.java.dir=src/java
+src.resources.dir=src/resources
 test.src.dir=test
