--- app/src/app/MainApp.java
@@ -1,21 +1,0 @@
-package app;
-
-import diario.departamentos.controllers.TableController;
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-public class MainApp extends Application {
-
-    @Override
-    public void start(Stage stage) throws Exception {
-        
-    }
-
-    public static void main(String[] args) {
-        launch(args);
-    }
-
-}
