--- app/src/diario/departamentos/resources/TabelaDepartamentos.fxml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.Cursor?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TableColumn?>
-<?import javafx.scene.control.TableView?>
-<?import javafx.scene.layout.StackPane?>
-<?import javafx.scene.shape.Rectangle?>
-<?import javafx.scene.text.Font?>
-
-<StackPane alignment="TOP_CENTER" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="720.0" prefWidth="1280.0" style="-fx-background-color: #E5E5E5;" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="diario.departamentos.controllers.TableController">
-   <children>
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="#29a58c" height="53.0" stroke="TRANSPARENT" strokeType="INSIDE" width="1280.0" />
-      <Label text="S.A. Admin" textFill="#fdfbf9" StackPane.alignment="TOP_LEFT">
-         <StackPane.margin>
-            <Insets left="10.0" top="8.0" />
-         </StackPane.margin>
-         <font>
-            <Font size="27.0" />
-         </font>
-      </Label>
-      <Button fx:id="adicionar" mnemonicParsing="false" onAction="#modalAdicionar" style="-fx-background-color: #29a58c;" text="ADICIONAR" textFill="#fdfbf9" StackPane.alignment="TOP_RIGHT">
-         <StackPane.margin>
-            <Insets right="240.0" top="150.0" />
-         </StackPane.margin>
-         <cursor>
-            <Cursor fx:constant="HAND" />
-         </cursor>
-         <font>
-            <Font size="16.0" />
-         </font>
-      </Button>
-      <Label fx:id="aviso" alignment="TOP_CENTER" contentDisplay="CENTER" style="-fx-background-color: red;" textFill="WHITE" StackPane.alignment="TOP_RIGHT">
-         <font>
-            <Font size="14.0" />
-         </font>
-         <StackPane.margin>
-            <Insets right="240.0" top="112.0" />
-         </StackPane.margin>
-      </Label>
-      <Label text="Departamentos" StackPane.alignment="TOP_LEFT">
-         <font>
-            <Font size="60.0" />
-         </font>
-         <StackPane.margin>
-            <Insets left="220.0" top="100.0" />
-         </StackPane.margin>
-      </Label>
-      <TableView fx:id="tabela" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="368.0" prefWidth="1036.0" StackPane.alignment="CENTER">
-        <columns>
-            <TableColumn fx:id="col_id" prefWidth="96.0" text="ID" />
-            <TableColumn fx:id="col_nome" minWidth="168.0" prefWidth="313.0" style="-fx-font-size: 16;" text="Nome" />
-          <TableColumn fx:id="col_campus" minWidth="296.0" prefWidth="313.0" style="-fx-font-size: 16;" text="Campus" />
-            <TableColumn fx:id="col_acoes" minWidth="308.0" prefWidth="313.0" style="-fx-font-size: 16;" text="Ações" />
-        </columns>
-         <StackPane.margin>
-            <Insets top="70.0" />
-         </StackPane.margin>
-      </TableView>
-   </children>
-</StackPane>
