--- app/src/diario/view/DepartamentosView.java
@@ -1,30 +1,0 @@
-package diario.view;
-
-import diario.departamentos.controllers.TableController;
-import javafx.application.Application;
-import static javafx.application.Application.launch;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-public class DepartamentosView extends Application {
-
-    @Override
-    public void start(Stage stage) throws Exception {
-        Parent root = FXMLLoader.load(getClass().getResource("/diario/departamentos/resources/TabelaDepartamentos.fxml"));
-
-        Scene scene = new Scene(root);
-        scene.getStylesheets().add(getClass().getResource("resources/styles.css").toExternalForm());
-        TableController t = new TableController();
-
-        stage.setTitle("Departamentos");
-        stage.setScene(scene);
-        stage.show();
-    }
-
-    public static void main(String[] args) {
-        launch(args);
-    }
-
-}
