+++ app/src/java/app/inicio/FXMLControllerInicio.java
@@ -1,0 +1,21 @@
+package app.inicio;
+
+import app.inicio.*;
+import java.net.URL;
+import java.util.ResourceBundle;
+import javafx.fxml.FXML;
+import javafx.fxml.Initializable;
+import javafx.scene.control.Label;
+
+public class FXMLControllerInicio implements Initializable {
+    
+    @FXML
+//    private Label label;
+    
+    @Override
+    public void initialize(URL url, ResourceBundle rb) {
+        String javaVersion = System.getProperty("java.version");
+        String javafxVersion = System.getProperty("javafx.version");
+//        label.setText("Hello, JavaFX " + javafxVersion + "\nRunning on Java " + javaVersion + ".");
+    }    
+}
