+++ app/src/resources/app/diario/departamentos/Departamentos.css
