+++ app/src/resources/app/diario/telaloginadm/EstiloTelaLoginAdm.css
@@ -1,0 +1,2 @@
+@import url('https://fonts.googleapis.com/css?family=Montserrat&display=swap');
+
