+++ app/src/resources/app/inicio/FXMLInicio.fxml
@@ -1,0 +1,98 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.*?>
+<?import javafx.scene.image.*?>
+<?import javafx.scene.text.*?>
+<?import javafx.scene.control.*?>
+<?import java.lang.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.StackPane?>
+
+<GridPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="720.0" prefWidth="1280.0" stylesheets="@EstiloInicio.css" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="app.inicio.FXMLControllerInicio">
+  <columnConstraints>
+    <ColumnConstraints hgrow="NEVER" minWidth="10.0" prefWidth="1280.0" />
+  </columnConstraints>
+  <rowConstraints>
+    <RowConstraints maxHeight="345.0" minHeight="10.0" prefHeight="100.0" vgrow="SOMETIMES" />
+    <RowConstraints maxHeight="589.0" minHeight="10.0" prefHeight="585.0" vgrow="SOMETIMES" />
+  </rowConstraints>
+   <children>
+      <AnchorPane prefHeight="200.0" prefWidth="200.0" GridPane.rowIndex="1">
+         <children>
+            <GridPane prefHeight="589.0" prefWidth="1280.0">
+              <columnConstraints>
+                <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+                <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+              </columnConstraints>
+              <rowConstraints>
+                <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+              </rowConstraints>
+               <children>
+                  <AnchorPane prefHeight="200.0" prefWidth="200.0">
+                     <children>
+                        <ImageView fitHeight="200.0" fitWidth="200.0" layoutX="120.0" layoutY="86.0">
+                           <image>
+                              <Image url="@../../imgs/Logo%20Diário.png" />
+                           </image>
+                        </ImageView>
+                        <Label layoutX="320.0" layoutY="141.0" styleClass="fontFamily" text="SISTEMA">
+                           <font>
+                              <Font size="30.0" />
+                           </font>
+                        </Label>
+                        <Label layoutX="320.0" layoutY="186.0" styleClass="fontFamily" text="DIÁRIO" textFill="#29a58c">
+                           <font>
+                              <Font size="40.0" />
+                           </font>
+                        </Label>
+                        <Button id="btnEntraSistemaDiario" layoutX="195.0" layoutY="371.0" mnemonicParsing="false" styleClass="backgroundBtn" text="ENTRAR" textFill="WHITE">
+                           <font>
+                              <Font size="24.0" />
+                           </font>
+                           <cursor>
+                              <Cursor fx:constant="HAND" />
+                           </cursor>
+                        </Button>
+                     </children></AnchorPane>
+                  <AnchorPane prefHeight="200.0" prefWidth="200.0" GridPane.columnIndex="1">
+                     <children>
+                        <ImageView fitHeight="200.0" fitWidth="200.0" layoutX="75.0" layoutY="87.0">
+                           <image>
+                              <Image url="@../../imgs/Logo%20Livraria.png" />
+                           </image>
+                        </ImageView>
+                        <Label layoutX="287.0" layoutY="187.0" styleClass="fontFamily" text="BIBLIOTECA" textFill="#29a58c">
+                           <font>
+                              <Font size="40.0" />
+                           </font>
+                        </Label>
+                        <Label layoutX="287.0" layoutY="142.0" styleClass="fontFamily" text="SISTEMA">
+                           <font>
+                              <Font size="30.0" />
+                           </font>
+                        </Label>
+                        <Button id="btnEntraSistemaBiblioteca" layoutX="195.0" layoutY="371.0" mnemonicParsing="false" styleClass="backgroundBtn" text="ENTRAR" textFill="WHITE">
+                           <font>
+                              <Font size="24.0" />
+                           </font>
+                           <cursor>
+                              <Cursor fx:constant="HAND" />
+                           </cursor>
+                        </Button>
+                     </children></AnchorPane>
+               </children>
+            </GridPane>
+         </children>
+      </AnchorPane>
+      <AnchorPane prefHeight="200.0" prefWidth="200.0">
+         <children>
+            <Label id="lblSistemaAcademico" alignment="CENTER" contentDisplay="CENTER" minWidth="1290.0" prefHeight="131.0" prefWidth="1280.0" styleClass="fontFamily" text="SISTEMA ACADÊMICO" textFill="WHITE">
+               <font>
+                  <Font size="48.0" />
+               </font>
+            </Label>
+         </children>
+      </AnchorPane>
+   </children>
+</GridPane>
