+++ pull_request_template.md
@@ -1,0 +1,16 @@
+# [Grupo X] Título
+
+* Mudanca 1
+* Mudanca 2
+* Mudanca 3
+
+## Cartões
+
+* [Nome do cartão](<--link do cartao (retirar as setas) -->)
+* [Nome do cartão](<--link do cartao (retirar as setas) -->)
+* [Nome do cartão](<--link do cartao (retirar as setas) -->)
+* [Nome do cartão](<--link do cartao (retirar as setas) -->)
+
+## Observações
+
+Sua observação aqui.
